### 2023-12-09 #6132 @editorialbot
[PRE REVIEW]: fABBA: A Python library for the fast symbolic approximation of time series
**Submitting author:** @chenxinye (<a href="http://orcid.org/0000-0003-1778-393X">Xinye Chen</a>)
**Repository:** https://github.com/nla-group/fABBA
**Branch with paper.md** (empty if default branch): master
**Version:** v1.1.0
