### 2023-12-11 97e1486
Update CITATION.bib
