+++ README.md
@@ -13,11 +13,11 @@
 [![Download Status](https://static.pepy.tech/badge/fABBA)](https://pypi.python.org/pypi/fABBA/)
 
 
-:sparkles: fABBA is a fast and accurate symbolic representation method for temporal data. 
+(https://zenodo.org/badge/DOI/10.5281/zenodo.10358352.svg)](https://doi.org/10.5281/zenodo.10358352): fABBA is a fast and accurate symbolic representation method for temporal data. 
 It is based on a polygonal chain approximation of the time series followed by an aggregation of the polygonal pieces into groups. 
 The aggregation process is sped up by sorting the polygonal pieces and exploiting early termination conditions. 
 In contrast to the ABBA method [S. Elsworth and S. Güttel, Data Mining and Knowledge Discovery, 34:1175-1200, 2020], fABBA avoids repeated within-cluster-sum-of-squares computations which reduces its computational complexity significantly.
-Furthermore, fABBA is fully tolerance-driven and does not require the number of time series symbols to be specified by the user. [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10358352.svg)](https://doi.org/10.5281/zenodo.10358352)
+Furthermore, fABBA is fully tolerance-driven and does not require the number of time series symbols to be specified by the user. [![DOI]
 
 ## :rocket: Install
 fABBA support Linux, Windows, and MacOS operating system [![Anaconda-Server Badge](https://anaconda.org/conda-forge/fabba/badges/platforms.svg)](https://anaconda.org/conda-forge/fabba). 
