### 2023-12-11 91ed9b4
Delete cpp directory
