+++ README.md
@@ -2,7 +2,7 @@
   fABBA
 </h1>
 <h3 align="left">
-  An efficient symbolic aggregate approximation for temporal data :dolphin:
+  An efficient symbolic aggregate approximation for temporal data  :dolphin:
 </h3>
 
 [![Build Status](https://app.travis-ci.com/nla-group/fABBA.svg?branch=master)](https://app.travis-ci.com/github/nla-group/fABBA)
