### 2023-12-22 f98cba6
Update aggmem.pyx -> fix svd
