+++ setup.py
@@ -26,7 +26,7 @@ class CustomBuildExtCommand(build_ext):
         
 setup_args = {'name':"fABBA",
         'packages':setuptools.find_packages(),
-        'version':"1.1.0",
+        'version':"1.1.1",
         'cmdclass': {'build_ext': CustomBuildExtCommand},
         'install_requires':["numpy>=1.3.0", "scipy>=0.7.0", 
                             "requests", "pandas", 
