+++ fABBA/separate/aggregation_cm.c
@@ -1892,7 +1892,7 @@ typedef npy_clongdouble __pyx_t_5numpy_clongdouble_t;
 typedef npy_cdouble __pyx_t_5numpy_complex_t;
 struct __pyx_opt_args_5fABBA_8separate_14aggregation_cm_aggregate;
 
-/* "fABBA/separate/aggregation_cm.pyx":53
+/* "fABBA/separate/aggregation_cm.pyx":29
  * 
  * 
  * cpdef aggregate(double[:,:] data, str sorting, double tol=0.5):             # <<<<<<<<<<<<<<
@@ -19483,7 +19483,7 @@ static CYTHON_INLINE NPY_DATETIMEUNIT __pyx_f_5numpy_get_datetime64_unit(PyObjec
   return __pyx_r;
 }
 
-/* "fABBA/separate/aggregation_cm.pyx":53
+/* "fABBA/separate/aggregation_cm.pyx":29
  * 
  * 
  * cpdef aggregate(double[:,:] data, str sorting, double tol=0.5):             # <<<<<<<<<<<<<<
@@ -19552,7 +19552,7 @@ static PyObject *__pyx_f_5fABBA_8separate_14aggregation_cm_aggregate(__Pyx_memvi
   int __pyx_clineno = 0;
   __Pyx_TraceFrameInit(__pyx_codeobj__11)
   __Pyx_RefNannySetupContext("aggregate", 1);
-  __Pyx_TraceCall("aggregate", __pyx_f[0], 53, 0, __PYX_ERR(0, 53, __pyx_L1_error));
+  __Pyx_TraceCall("aggregate", __pyx_f[0], 29, 0, __PYX_ERR(0, 29, __pyx_L1_error));
   if (__pyx_optional_args) {
     if (__pyx_optional_args->__pyx_n > 0) {
       __pyx_v_tol = __pyx_optional_args->tol;
@@ -19560,149 +19560,149 @@ static PyObject *__pyx_f_5fABBA_8separate_14aggregation_cm_aggregate(__Pyx_memvi
   }
   __PYX_INC_MEMVIEW(&__pyx_v_data, 1);
 
-  /* "fABBA/separate/aggregation_cm.pyx":80
+  /* "fABBA/separate/aggregation_cm.pyx":56
  *         distance computation calculations
  *     """
  *     cdef Py_ssize_t fdim = data.shape[1] # feature dimension             # <<<<<<<<<<<<<<
  *     cdef Py_ssize_t len_ind = data.shape[0] # size of data
  *     cdef double[:] sort_vals
  */
-  __Pyx_TraceLine(80,0,__PYX_ERR(0, 80, __pyx_L1_error))
+  __Pyx_TraceLine(56,0,__PYX_ERR(0, 56, __pyx_L1_error))
   __pyx_v_fdim = (__pyx_v_data.shape[1]);
 
-  /* "fABBA/separate/aggregation_cm.pyx":81
+  /* "fABBA/separate/aggregation_cm.pyx":57
  *     """
  *     cdef Py_ssize_t fdim = data.shape[1] # feature dimension
  *     cdef Py_ssize_t len_ind = data.shape[0] # size of data             # <<<<<<<<<<<<<<
[…]
