+++ fABBA/jabba/aggmem.c
@@ -19865,7 +19865,7 @@ static PyObject *__pyx_f_5fABBA_5jabba_6aggmem_aggregate(__Pyx_memviewslice __py
  *         # cdata = data - data.mean(axis=0) -- already done in the clustering.fit_transform
  *         cdata = data - np.mean(data, axis=0)             # <<<<<<<<<<<<<<
  *         if data.shape[1]>1:
- *             U1, s1, _ = svds(cdata, k=1, return_singular_vectors=True)
+ *             U1, s1, _ = svds(cdata.base, k=1, return_singular_vectors=True)
  */
   __Pyx_TraceLine(77,0,__PYX_ERR(0, 77, __pyx_L1_error))
   /*else*/ {
@@ -19906,7 +19906,7 @@ static PyObject *__pyx_f_5fABBA_5jabba_6aggmem_aggregate(__Pyx_memviewslice __py
  *         # cdata = data - data.mean(axis=0) -- already done in the clustering.fit_transform
  *         cdata = data - np.mean(data, axis=0)
  *         if data.shape[1]>1:             # <<<<<<<<<<<<<<
- *             U1, s1, _ = svds(cdata, k=1, return_singular_vectors=True)
+ *             U1, s1, _ = svds(cdata.base, k=1, return_singular_vectors=True)
  *             sort_vals = U1[:,0]*s1[0]
  */
     __Pyx_TraceLine(78,0,__PYX_ERR(0, 78, __pyx_L1_error))
@@ -19916,7 +19916,7 @@ static PyObject *__pyx_f_5fABBA_5jabba_6aggmem_aggregate(__Pyx_memviewslice __py
       /* "fABBA/jabba/aggmem.pyx":79
  *         cdata = data - np.mean(data, axis=0)
  *         if data.shape[1]>1:
- *             U1, s1, _ = svds(cdata, k=1, return_singular_vectors=True)             # <<<<<<<<<<<<<<
+ *             U1, s1, _ = svds(cdata.base, k=1, return_singular_vectors=True)             # <<<<<<<<<<<<<<
  *             sort_vals = U1[:,0]*s1[0]
  *         else:
  */
@@ -19925,20 +19925,23 @@ static PyObject *__pyx_f_5fABBA_5jabba_6aggmem_aggregate(__Pyx_memviewslice __py
       __Pyx_GOTREF(__pyx_t_3);
       __pyx_t_1 = __pyx_memoryview_fromslice(__pyx_v_cdata, 2, (PyObject *(*)(char *)) __pyx_memview_get_double, (int (*)(char *, PyObject *)) __pyx_memview_set_double, 0);; if (unlikely(!__pyx_t_1)) __PYX_ERR(0, 79, __pyx_L1_error)
       __Pyx_GOTREF(__pyx_t_1);
-      __pyx_t_4 = PyTuple_New(1); if (unlikely(!__pyx_t_4)) __PYX_ERR(0, 79, __pyx_L1_error)
[…]
