### 2023-12-22 5212a60
Update fabba_agg_c.pyx
