### 2024-01-22 #6132 @Karangupta1994
Hi Oskar,

Sure, I will be happy to review. Please let me know how to proceed.

Thank you,
Karan

On Mon, Jan 22, 2024 at 4:27 AM Oskar Laverny ***@***.***>
wrote:

> 👋 @allie-tatarian <https://github.com/allie-tatarian> @@Karangupta1994
> <https://github.com/Karangupta1994> @djmannion
> <https://github.com/djmannion> @dgerosa <https://github.com/dgerosa>
> @wmvanvliet <https://github.com/wmvanvliet>, would any of you be willing
> to review this submission for JOSS? We carry out our checklist-driven
> reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
