### 2024-02-01 #6294 @allie-tatarian
Hi all, I'm not sure if this is the best place to this but I don't think it makes sense to post this as an issue because I think it's a problem on my end. When I try to use fABBA, I am consistently getting a "This installation is not using Cython" warning even though I do have Cython installed. @chenxinye is this something you have seen before? I cannot for the life of me figure out how to troubleshoot this. The fABBA library is still mostly working for me, but sometimes I get different outputs than the examples.
