### 2024-02-01 #6294 @lrnv
@allie-tatarian Different output than the examples surely looks like an issue (unless the examples are explicitely random ?). Problem in the installation can also be reported as issues.

@Karangupta1994 & @allie-tatarian, to guide you in the review, there is an automated checklist that you have to produce by commenting in this thread :

`@editorialbot generate my checklist`

I think the standard checkpionts includes stuff such as the need for descriptive installation guidelines if some stuff are "non-standard"
