+++ paper/paper.bib
@@ -1,21 +1,23 @@
 @article{EG19b,
-  title   = {{ABBA: adaptive Brownian bridge-based symbolic aggregation of time series}},
-  author  = {Elsworth, Steven and G\"{u}ttel, Stefan},
-  year    = {2020},
-  volume  = {34},
-  pages   = {1175--1200},
-  journal = {Data Mining and Knowledge Discovery}
+title   = {{ABBA: adaptive Brownian bridge-based symbolic aggregation of time series}},
+author  = {Elsworth, Steven and G\"{u}ttel, Stefan},
+year    = {2020},
+volume  = {34},
+doi = {10.1007/s10618-020-00689-6}, 
+pages   = {1175--1200},
+journal = {Data Mining and Knowledge Discovery}
 }
 
 @ARTICLE{1056489,
-  author={Lloyd, Stuart P.},
-  journal={Transactions on Information Theory}, 
-  publisher={IEEE},
-  title={Least squares quantization in {PCM}}, 
-  year={1982},
-  volume={28},
-  pages={129--137}
- }
+author={Lloyd, Stuart P.},
+journal={Transactions on Information Theory}, 
+publisher={IEEE},
+title={Least squares quantization in {PCM}}, 
+year={1982},
+volume={28},
+doi={10.1109/tit.1982.1056489},
+pages={129--137}
+}
 
 @inproceedings{SAX03,
 author = {Lin, Jessica and Keogh, Eamonn and Lonardi, Stefano and Chiu, Bill},
@@ -24,19 +26,21 @@ year = {2003},
 publisher = {ACM},
 booktitle = {Proceedings of the 8th ACM SIGMOD Workshop on Research Issues in Data Mining and Knowledge Discovery},
 pages = {2--11},
+doi={10.1145/882082.882086},
 numpages = {10},
 series = {DMKD '03}
 }
 
 
 @article{CG22a,
-  title   = {An efficient aggregation method for the symbolic representation of temporal data},
-  author  = {Chen, Xinye and G\"{u}ttel, Stefan},
-  year    = {2023},
-  volume  = {17},
-  number  = {1},
-  pages   = {1--22},
-  journal = {ACM Transactions on Knowledge Discovery from Data}
+title   = {An efficient aggregation method for the symbolic representation of temporal data},
+author  = {Chen, Xinye and G\"{u}ttel, Stefan},
+year    = {2023},
+volume  = {17},
+number  = {1},
+doi = {10.1145/3532622}, 
+pages   = {1--22},
+journal = {ACM Transactions on Knowledge Discovery from Data}
 }
 
 @article{TAKTAK2024102294,
@@ -45,6 +49,7 @@ journal = {Information Systems},
 volume = {120},
 pages = {102294},
 year = {2024},
+doi = {10.1016/j.is.2023.102294},
 author = {Mariem Taktak and Hela Ltifi and Mounir Ben Ayed},
 }
 
@@ -53,53 +58,57 @@ author = {Nguyen, Thach Le and Ifrim, Georgiana},
 title = {Fast Time Series Classification with Random Symbolic Subsequences},
 year = {2023},
 publisher = {Springer},
+doi={10.1007/978-3-031-24378-3_4},
 booktitle = {Advanced Analytics and Learning on Temporal Data: 7th ECML PKDD Workshop},
 pages = {50--65},
 numpages = {16},
 }
 
 @article{lin2007experiencing,
-  author = {Lin, Jessica and Keogh, Eamonn and Wei, Li and Lonardi, Stefano},
-  journal = {Data Mining and Knowledge Discovery},
-  number = 2,
-  pages = {107--144},
-  publisher = {Springer},
-  title = {Experiencing {SAX}: a novel symbolic representation of time series},
[…]
