+++ README.md
@@ -119,7 +119,7 @@ from fABBA import ABBAbase
 
 ts = [np.sin(0.05*i) for i in range(1000)]         # original time series
 #  specifies 5 symbols using kmeans clustering
-kmeans = KMeans(n_clusters=5, random_state=0, init='k-means++', verbose=0)     
+kmeans = KMeans(n_clusters=5, random_state=0, init='k-means++', n_init='auto', verbose=0)     
 abba = ABBAbase(tol=0.1, scl=1, clustering=kmeans)
 string = abba.fit_transform(ts)                    # string representation of the time series
 print(string)                                      # prints BbAaAaAaAaAaAaAaC
