### 2024-02-01 #6294 @chenxinye
@allie-tatarian

I just checked, found some output in README.md is consistent with the real output of software. I now corrected it! Please check.
