### 2024-02-01 #6294 @chenxinye
Thanks for pointing the issues.

For the Cython issue, @lrnv @allie-tatarian, some users reported before, most of them have this issue since they machine have firewall or without VC++ installed. Could you please provide your machine specifications for me?

I checked my three machines, all worked properly. I guess you use Windows without VC++ installed?
