### 2024-02-01 #6294 @chenxinye
You would need to install fABBA using:

``pip install fABBA``
