### 2024-02-02 #6294 @allie-tatarian
Thanks @chenxinye, like I mentioned in the issue I did install using pip. It is working fine on my personal computer, so I think the issue is that the security on my work computer is not meshing well with cython. I wonder if it's worth putting a note about that in the docs for fABBA - I do think the issue is with cython, not fABBA, but looking around online it looks like I'm not the only one who has run into this issue and I haven't seen anyone spell out the solution as concisely as you did here!
