+++ doc/source/quickstart.rst
@@ -57,7 +57,7 @@ and uninstall it with
 
 .. admonition:: Note
    
-   Some machine may raise the Cython issue, most of them happens since their machine have security issue - e.g., firewall - or without VC++ installed. 
+   Some machine may raise the Cython issue, most of them happens since their machine have security issue - e.g., firewall - or without VC++ installed. Pleasee contact developers team for further support if you cannot address the Cython related issues. 
 
 Installing `fABBA` from the `conda-forge` channel can also be achieved by adding `conda-forge` to your channels with:
 
