### 2024-02-02 #6294 @chenxinye
@allie-tatarian  Just added in https://fabba.readthedocs.io/en/latest/quickstart.html ! Many thanks
