### 2024-02-06 #6294 @lrnv
@allie-tatarian I'm confused, what is the difference between fABBA (the software we review here) and JABBA ? Functionalities of the reviewed software need to be checked up, even if they are not mentionned in the paper itself. e.g., you can check out the documentation, see if it works correctly, etc..

@chenxinye is JABBA part of the submitted work here ?
