### 2024-02-06 fceb372
Update multivariate.rst
