+++ doc/source/multivariate.rst
@@ -2,7 +2,7 @@ Multivariate time series symbolization
 ======================================
 
 
-Here we domonstrate how to use ``fABBA`` to symbolize multivariate (same applies to multiple univariate time series) with consistent symbols.
+Here we domonstrate how to use ``fABBA`` to symbolize multivariate (same applies to multiple univariate time series) with consistent symbols. After downloading the `UEA time series dataset <https://www.timeseriesclassification.com/>`_ in corresponding folder, you can run JABBA following the example below:
 
 
 .. code:: python
