### 2024-02-06 #6294 @chenxinye
Sorry for the confusion @lrnv @allie-tatarian

JABBA is an extension of ABBA/fABBA towards multiple or multivariate time series.

@allie-tatarian  Which issue did you come across,
did you follow this:  https://fabba.readthedocs.io/en/latest/multivariate.html?
