### 2024-02-06 #6294 @chenxinye
The method JABBA also works for Before running the example, you should download the UEA data: https://timeseriesclassification.com/

Or you can simply generate multivariate time series yourself via: ``np.random.randn(50, 30)``
