### 2024-02-06 #6294 @allie-tatarian
Thank you both for the clarification! Yes, @chenxinye, that is the code I was following. I downloaded the BasicMotions dataset, copied the code, and the only thing I changed was changing the directory in the first line to an absolute directory pointing to the dataset. When I try to run it, I get the error: `AttributeError: module 'os' has no attribute 'sched_getaffinity'` which gets called at line 28, when the code is trying to call fit_transform(). When I looked up the error, it said that sched_getaffinity only works on Linux machines, not Windows. The full error message is:
```
Traceback (most recent call last):
  File "c:\Users\willi\Documents\program_practice\fabba\jabba_docs.py", line 28, in <module>
    symbols_series = jabba1.fit_transform(mts)
                     ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 262, in fit_transform
    self.fit(series, n_jobs=n_jobs, alphabet_set=alphabet_set)
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 289, in fit
    self.pieces = self.parallel_compress(series, n_jobs=n_jobs)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 317, in parallel_compress
    n_jobs = self.n_jobs_init(n_jobs, _max=len_ts)
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 655, in n_jobs_init
    n_jobs = len(os.sched_getaffinity(0))
                 ^^^^^^^^^^^^^^^^^^^^
AttributeError: module 'os' has no attribute 'sched_getaffinity'
```
