### 2024-02-06 #6294 @allie-tatarian
A separate but related issue I ran into when trying to run the code that is in the README in the [JABBA section(https://github.com/nla-group/fABBA?tab=readme-ov-file#star-for-multiple-time-series-data-transform) is that that code includes the lines:
``` python
from fABBA import loadData
train, test = loadData()
```
I can't find any documentation, either in the README or the documentation, about how loadData() works. The image compression section of the README has a similar issue - it introduces load_images(), but I can't find anything about how to use load_images in the README or the docs.
