### 2024-02-07 886d1ed
Update multivariate.rst
