+++ fABBA.egg-info/SOURCES.txt
@@ -22,6 +22,7 @@ fABBA/extmod/fabba_agg_cm.pyx
 fABBA/extmod/inverse_tc.pyx
 fABBA/jabba/__init__.py
 fABBA/jabba/agg.py
+fABBA/jabba/aggc.pyx
 fABBA/jabba/aggmem.pyx
 fABBA/jabba/comp.py
 fABBA/jabba/compmem.pyx

[…]
