+++ build/lib.linux-x86_64-3.11/fABBA/__init__.py
@@ -10,8 +10,8 @@ except (ModuleNotFoundError, ValueError):
     from .separate.aggregation import aggregate
     
     
-__version__ = '1.1.5'
-from .load_datasets import load_images
+__version__ = '1.1.7'
+from .load_datasets import load_images, loadData
 from .fabba import (image_compress, image_decompress, ABBAbase, ABBA,
                                       get_patches, patched_reconstruction, fABBA,
                                       symbolsAssign, fillna)
@@ -24,4 +24,3 @@ from .jabba.jabba import fastABBA, JABBA
 from .jabba import jabba
 from .jabba import preprocessing
 from .jabba.fkmns import sampledKMeansInter as fkmeans
-from .jabba.loadData import *
