+++ doc/source/api_reference.rst
@@ -28,7 +28,7 @@ JABBA
 ``JABBA`` is the API of the symbolic representation transformation for univariate time series, multivariate (rep., multiple univariate) time series, which allows for the combination of ABBA method with various clustering techniques. 
 
 
-.. autoclass:: fABBA.jabba.jabba.JABBA
+.. autoclass:: fABBA.JABBA
    :members:
    
 
