### 2024-02-07 #6294 @chenxinye
Hi @allie-tatarian @lrnv , thanks for mentioning that, I just add documentation as well as docstring for the ``loadData()``.

The updated document can be found in https://fabba.readthedocs.io/en/latest/api_reference.html.

Also I fix the JABBA issue in Windows, many thanks for raising this issue!

I will also continue to work on the improvements of the documentation.
