### 2024-02-07 #6294 @allie-tatarian
@chenxinye Thanks for the quick update! I tried the code in the documentation and it worked perfectly. I also tried the code with `loadData()` from the README, and it works when I use "Beef" but not "BasicMotions." I checked the directory that the data should have been in, and the "Beef" data was there but not the BasicMotions folder. I added the BasicMotions folder I had downloaded, and I got this error:
```
  File "c:\Users\willi\Documents\program_practice\fabba\jabba_readme.py", line 5, in <module>
    train, test = loadData(name="BasicMotions")
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\load_datasets.py", line 47, in loadData
    return train, test
                  ^^^^
UnboundLocalError: cannot access local variable 'test' where it is not associated with a value
```

Also I can see now the documentation for `load_images()` too, thanks for clarifying how it works! I hadn't tried that code yet, that will be the next thing I do.
