### 2024-02-07 #6294 @allie-tatarian
Quick update: The image compression/decompression code worked great, so the only issue I have left is when using `loadData(name="BasicMotions")`.
