+++ setup.py
@@ -34,7 +34,7 @@ setup_args = {'name':"fABBA",
                             "joblib>=1.1.1",
                             "matplotlib"],
         'packages':{"fABBA", "fABBA.extmod", "fABBA.separate", "fABBA.jabba"},
-        'package_data':{"fABBA": ["jabba/data/*.npy"]},
+        'package_data':{"fABBA": ["jabba/data/*.npy", "jabba/data/*.arff"]},
         # 'include_dirs':[numpy.get_include()],
         'long_description':long_description,
         'author':"Xinye Chen, Stefan Güttel",
