### 2024-02-08 d3fbe7d
Update multivariate.rst
