### 2024-02-08 d532e7d
Update multivariate.rst
