+++ doc/source/quickstart.rst
@@ -1,5 +1,5 @@
 
-Get Started with fABBA
+Get Started
 ======================================
 
 
