### 2024-02-08 8df376e
Update multivariate.rst
