### 2024-02-09 #6294 @allie-tatarian
Thanks @chenxinye, now when I try to use `loadData(name="BasicMotions")`, I get this error (I do NOT get an error when I try to run the code using `loadData(name="Beef")`...again, this is the [JABBA example code in the README](https://github.com/nla-group/fABBA?tab=readme-ov-file#star-for-multiple-time-series-data-transform)):

```
  File "c:\Users\willi\Documents\program_practice\fabba\jabba_readme.py", line 8, in <module>
    symbols = jabba.fit_transform(train)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 267, in fit_transform
    self.fit(series, n_jobs=n_jobs, alphabet_set=alphabet_set)
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 294, in fit
    self.pieces = self.parallel_compress(series, n_jobs=n_jobs)
                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\site-packages\fABBA\jabba\jabba.py", line 367, in parallel_compress
    pieces = [p.get() for p in pieces]
              ^^^^^^^
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\multiprocessing\pool.py", line 774, in get
    raise self._value
  File "C:\Users\willi\AppData\Local\Programs\Python\Python312\Lib\multiprocessing\pool.py", line 125, in worker
    result = (True, func(*args, **kwds))
                    ^^^^^^^^^^^^^^^^^^^
  File "fABBA\\jabba\\compmem.pyx", line 10, in fABBA.jabba.compmem.compress
ValueError: Buffer has wrong number of dimensions (expected 1, got 2)
```
