+++ README.md
@@ -140,9 +140,13 @@ train, test = loadData()
 Built in ``JABBA`` provide parameter of ``init`` for the specification of ABBA methods, if set ``agg``, then it will automatically turn to fABBA method, and if set it to ``k-means``, it will turn to ABBA method automatically. Use ``JABBA`` object to fit and symbolize the train set via API ``fit_transform``, and reconstruct the time series from the symbolic representation simply by
 ``` Python
 jabba = JABBA(tol=0.0005, init='agg', verbose=1)
-symbols = jabba.fit_transform(train)
+symbols = jabba.fit_transform(train) 
 reconst = jabba.inverse_transform(symbols)
 ```
+
+Note: JABBA is used to process multiple time series as well as multivariate time series, so the input should be ensured to be 2-dimensional, e.g., when loading UCI dataset, e.g., ``Beef``, use symbols = jabba.fit_transform(train), when loading UEA dataset, use symbols = jabba.fit_transform(train[0]).
+
+
 
 For the out-of-sample data, use the function ``transform`` to symbolize the test time series, and reconstruct the symbolization via function  ``inverse_transform``, the code illustration is as follows: 
 ``` Python
