### 2024-02-09 #6294 @chenxinye
Hi @allie-tatarian, Thank you, I missed that. JABBA is used to process the 2-dimensional time series, the 3-dimensional one have not yet developed.

So the proper use is

```
from fABBA import JABBA
from fABBA import loadData
train, test = loadDataname="BasicMotions"()
jabba = JABBA(tol=0.0005, init='agg', verbose=1)
symbols = jabba.fit_transform(train[0])
reconst = jabba.inverse_transform(symbols)
```

I have added note in the ([[JABBA example code in the README](https://github.com/nla-group/fABBA?tab=readme-ov-file#star-for-multiple-time-series-data-transform)](https://github.com/nla-group/fABBA?tab=readme-ov-file#star-for-multiple-time-series-data-transform))
