### 2024-02-09 2e61283
Update multivariate.rst
