### 2024-02-09 #6294 @chenxinye
@allie-tatarian

Exactly, thanks for mentioning it, I already added note in the code example for the ``transform()`` method.

> Ok, great, thanks for the explanation @chenxinye! I also noticed that for the last two lines:
>
> ```python
> test_symbols, start_set = jabba.transform(test)
> test_reconst = jabba.inverse_transform(test_symbols, start_set)
> ```
>
> I also had to change `jabba.transform(test)` to `jabba.transform(test[0])` to get it to work using BasicMotions.
>
> Thanks for all of your help and quick updates!
