### 2024-02-29 #6294 @chenxinye
I just checked my DOI 10.5555/1953048.2078195 is correct, not sure why it reminds me invalid DOI.
