### 2024-02-29 #6294 @chenxinye
> @chenxinye No I had an e-mail exchange with @Karangupta1994 and It looks like he's still on board. Why are you saying his accoutn is frozen ?

Sorry, I apologized the mistake! I only saw the account activity.
