### 2024-02-29 #6294 @chenxinye
Hi @lrnv  it seems like https://www.jmlr.org/papers/v12/pedregosa11a.html does not have approprite DOI, how do I deal with it? Many thanks!
