+++ paper/paper.bib
@@ -77,18 +77,15 @@ year = 2007
 }
 
 @article{scikit-learn,
-title={Scikit-learn: Machine Learning in {P}ython},
-author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
-       and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
-       and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
-       Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
-journal={Journal of Machine Learning Research},
-volume={12},
-doi={10.5555/1953048.2078195},
-pages={2825--2830},
-year={2011}
+  author  = {Fabian Pedregosa and Ga{{\"e}}l Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and {{\'E}}douard Duchesnay},
+  title   = {Scikit-learn: Machine Learning in Python},
+  journal = {Journal of Machine Learning Research},
+  year    = {2011},
+  volume  = {12},
+  number  = {85},
+  pages   = {2825--2830},
+  url     = {http://jmlr.org/papers/v12/pedregosa11a.html}
 }
-
 
 @INPROCEEDINGS{9935005,
 author={Gogineni, Kailash and Derasari, Preet and Venkataramani, Guru},
