### 2024-02-29 #6294 @chenxinye
> I think you just have to curate your bib file and remove the DOi field. At the link you provided, they give the following bib :
>
> ```
> @article{JMLR:v12:pedregosa11a,
>   author  = {Fabian Pedregosa and Ga{{\"e}}l Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and {{\'E}}douard Duchesnay},
>   title   = {Scikit-learn: Machine Learning in Python},
>   journal = {Journal of Machine Learning Research},
>   year    = {2011},
>   volume  = {12},
>   number  = {85},
>   pages   = {2825--2830},
>   url     = {http://jmlr.org/papers/v12/pedregosa11a.html}
> }
> ```
>
> without a doi field.

Many thanks, now sorted!
