### 2024-03-24 #6294 @lrnv
@chenxinye Is there somewhere is your documentation a point htat would address @Karangupta1994 third remark about compatibility with python/np/pd versions ? If no, would you accept to add such a statement ?

After that, I think we all agree that this is very good work on your side and we'll start the acceptation and publication process. Congratz!
