### 2024-03-26 #6294 @chenxinye
Hi @Karangupta1994
Many thanks for the insightful suggestions.

For (1), thanks for the great advice, I add some data source to https://github.com/nla-group/fABBA/blob/master/README.md#checkered_flag-examples
The comment is: To play fABBA further with real datasets, we recommend users start with [UCI Repository](https://archive.ics.uci.edu/datasets?skip=0&take=10&sort=desc&orderBy=NumHits&search=&Types=Time-Series) and [UCR Archive](https://www.cs.ucr.edu/%7Eeamonn/time_series_data_2018/).

For (2) I think the paper of fABBA (see https://dl.acm.org/doi/full/10.1145/3532622), which provides full description of evalution.

For (3) I just updated in REDEAME.md and documentation (https://fabba.readthedocs.io/en/latest/quickstart.html) to specify the version of dependencies required.

Best regards!
