### 2024-03-27 #6294 @chenxinye
Hi @lrnv

* The author information is correct.
* The submitted version of the software is PyPI 1.2.1, changes in the repo is updated
* DOI: https://doi.org/10.5281/zenodo.10885652
* Completed
* Completed

Many thanks!
