### 2024-03-27 #6294 @lrnv
@chenxinye Licence on github says BSD 3-Clause while is says CC4.0 on zenodo ? Could you fix one or the other ?
