### 2024-03-27 #6294 @chenxinye
Hi @lrnv I would need to add some comments in Acknowledgement, can I?

``Stefan Güttel acknowledges a Royal Society Industry Fellowship IF/R1/231032. Xinye Chen is supported by the European Union (ERC, inEXASCALE, 101075632). Views and opinions expressed are those of the authors only and do not necessarily reflect those of the European Union or the European Research Council. Neither the European Union nor the granting authority can be held responsible for them.''

Can I?
